Moved fuzzy_match_timezone to bios/misc.py.
